requirements.txt
# claudey-tr dependencies
# Install with: pip install -r requirements.txt

# ── Angel One broker API ──────────────────────────────────────────────────────
smartapi-python>=1.3.4          # Angel One SmartAPI client
pyotp>=2.9.0                    # TOTP generation for daily 2FA
websocket-client>=1.7.0         # Angel One WebSocket feed

# ── Anthropic / Claude ────────────────────────────────────────────────────────
anthropic>=0.40.0               # Claude API SDK
fastmcp>=2.0.0                  # MCP server (Phase 4 — Claude tool integration)

# ── Data & indicators ─────────────────────────────────────────────────────────
pandas>=2.1.0                   # candle data, rolling indicators
numpy>=1.26.0                   # numerical operations

# ── Scheduling & async ────────────────────────────────────────────────────────
APScheduler>=3.10.4             # daily login/logout scheduling
aiosqlite>=0.20.0               # async SQLite (Phase 4+)

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0            # .env loading
pytz>=2024.1                    # IST timezone handling
requests>=2.31.0                # HTTP calls (news feeds, IP check)
feedparser>=6.0.11              # RSS parsing (handles malformed real-world feeds)
lxml>=5.0.0                     # feedparser uses lxml with recover=True for malformed XML

# ── Offline backtesting (scripts/ only — not used in live trading) ────────────
yfinance>=0.2.40                # USDINR=X historical data for offline backtest

# ── Testing ───────────────────────────────────────────────────────────────────
pytest>=7.4.0
pytest-asyncio>=0.23.0
